Switch to wabt.js
It's somewhat easier to use from JavaScript, makes wabt a regular npm dependency, and, what's even more important, is more frequently updated than WABT pre-built binaries.

It's not yet updated to fix #22 for us, but should be on the next nightly build - feel free to either merge this now or keep open until it gets updated.