Fix threads detection on nodejs
We have a separate file for threads detection on nodejs,
threads/index-node.js. That means all the detectors' index.js should
work on both browser and node.

We generate new files in dist for node env, suffixed with `-node`.
Update package.json to point to the correct files:

- main and module is used by nodejs, it points to `index-node.js`
- browser is used for browser, so points to the `index.js`
- unpkg is used by browser (at least in the examples in README) so point
to `index.js` too

The rollup config is now a nested generation:
- for each env (browser and node),
- for each format (esm, cjs, umd),
- generate files

However the dist folder is still only 1 level deep, indexed by the
format.