Fix SIMD detection
A bunch of opcodes changed in the final version of SIMD (IIUC, for the last time): https://chromium.googlesource.com/v8/v8.git/+/c3f9d3cd8eb6a7a41c27f2c783c11e4cffb21a2d

However, our test has been checking for an opcode that didn't change between versions, so it happens to report success on engines with both old and new SIMD implementations, apps that depend on wasm-feature-detect can use that status to load SIMD version of Wasm, but then it happens to break depending on toolchain + engine combo.

This PR updates SIMD test to check against the new instruction's opcode as well, so that the test passes only on the final version of the spec.