Some prettier-related fixes
This applies prettier to README.md.ejs, replacing tabs with spaces there.

Also, it seems that fastGlob has some limitation that prevents it from
correctly handling the glob pattern, so that prettier is not applied to
README.md or other files in the root directory of the package. I worked
around this limitation by duplicationg the list of file endings.